Move WFSLayerPermissionsStore to control-base where it's used
