+++ control-base/src/main/java/fi/nls/oskari/control/layer/GetLayerIds.java
@@ -1,7 +1,7 @@
 package fi.nls.oskari.control.layer;
 
 import fi.mml.map.mapwindow.util.OskariLayerWorker;
-import fi.mml.portti.domain.permissions.WFSLayerPermissionsStore;
+import fi.nls.oskari.util.WFSLayerPermissionsStore;
 import fi.nls.oskari.annotation.OskariActionRoute;
 import fi.nls.oskari.control.ActionException;
 import fi.nls.oskari.control.ActionHandler;
@@ -14,11 +14,9 @@ import fi.nls.oskari.util.JSONHelper;
 import fi.nls.oskari.util.PropertyUtil;
 import fi.nls.oskari.util.ResponseHelper;
 import org.json.JSONArray;
-import org.json.JSONObject;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 /**
